fix: ouster_ros ring data type changed to uint16_t to match ros2 driver

+++ src/preprocess.h
@@ -58,7 +58,7 @@ namespace ouster_ros {
       float intensity;
       uint32_t t;
       uint16_t reflectivity;
-      uint8_t  ring;
+      uint16_t  ring;
       uint16_t ambient;
       uint32_t range;
       EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -74,7 +74,7 @@ POINT_CLOUD_REGISTER_POINT_STRUCT(ouster_ros::Point,
     // use std::uint32_t to avoid conflicting with pcl::uint32_t
     (std::uint32_t, t, t)
     (std::uint16_t, reflectivity, reflectivity)
-    (std::uint8_t, ring, ring)
+    (std::uint16_t, ring, ring)
     (std::uint16_t, ambient, ambient)
     (std::uint32_t, range, range)
 )
